Support swiping back
Override the onBackPress callback, which by default brings the application to the background and instead tell servo to go back one page.
Users can invoke this by swiping from the left (or right) edge to the middle. There is no equivalent callback / gesture to go forward. In the default browser swiping from the right side to the middle also invokes the goBack callback.